==> 1_Webscraping_Datasets_And_Appending.ipynb <==
# Webscraping the MTA dataset and appending multiple dataframes

This Notebook shows the code used in the MTA data analysis.

Group: Danny, Navina, John, Kevin

Import Libraries

from bs4 import BeautifulSoup
import requests
import pprint


Create an array of urls

url = "http://web.mta.info/developers/turnstile.html"
page = requests.get(url).text
soup = BeautifulSoup(page, features="html5lib")
urls = []

for a in soup.find_all('a', href=True):
    urls.append(["http://web.mta.info/developers/" + a['href']])

urls = urls[35:44]

pprint.pprint(urls)

If the urls are saved to a file, we can import them with the following code:

'''
with open('MTA_urls.txt', 'w') as file_handler:
    for item in urls:
        file_handler.write("{}\n".format(item[0]))
'''

Read the data and append together

for url in urls:
    page = requests.get(url).text
    new_df = pd.read_csv(url, sep='\n')
    MTA_df = pd.concat([MTA_df, new_df])
    
print(MTA_df.describe())

Uncomment to save the data to a file. It is ~200Mb in size.

#MTA_df.to_csv(r'MTA_all_data.csv', index=None, sep=' ', mode='a')

# Warning!! Problematic commas in the file, please run the following on the command line

In terminal: sed -i -e 's/"$//'  MTA_all_data.csv
