==> 4_MTA_Turnstile_Data_Wrangling.ipynb <==
# NYC MTA Turnstile Dataset Cleanup and Merging Data from Different Sources

from IPython.display import display
import warnings
import calendar
warnings.filterwarnings('ignore')
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns
sns.set_style("whitegrid")
%pylab inline

# Process and dump a single file
def delta_processor(filename):
    df = pd.read_csv(filename)
    
    #Cleanup column names - remove spaces
    df.rename(columns=lambda x: x.strip(), inplace=True)

    # Cleaning datetime
    print("Cleaning up date time")
    df.loc[:,'DATETIME'] = pd.to_datetime(df['DATE'] + ' ' + df['TIME'])
    df['DAY'] = df['DATETIME']
    df['DAY'] = df['DAY'].apply(lambda x : calendar.day_name[x.weekday()])
    
    # diff entry and exit and update dataframe
    print('Calculate running delta for entry/exit time')
    entries_diff = df['ENTRIES'].diff()
    entries_diff = entries_diff.clip_lower(0)
    df['ENTRIES_DELTA'] = entries_diff
    exits_diff = df['EXITS'].diff()
    exits_diff = exits_diff.clip_lower(0)
    df['EXITS_DELTA'] = exits_diff
    df.fillna(0, inplace=True)
    
    # First entry/exit delta for each day in each station and SCP will be incorrect. Drop these results
    # Next, there are chances for spurious deltas if data is not correct, ignore such outliers
    print('Removing spurious records')
    df = df.groupby(['STATION', 'SCP', 'DATE'], group_keys=False).apply(lambda group: group.iloc[1:])
    df = df[df['ENTRIES_DELTA'] < 10000]
    df = df[df['EXITS_DELTA'] < 10000]
    
    #TODO: Write to CSV
    # 'abc.csv' -> 'abc_delta_processed.csv'
    return df 
    

turnstile_df = delta_processor('MTA_all_data.csv')

turnstile_df[[ 'UNIT','DIVISION', 'STATION', 'DATE', 'TIME','DATETIME', 'DAY', 'ENTRIES_DELTA', 'EXITS_DELTA']].head()

#  out to csv after preliminary cleaning and diff generation
## turnstile_df.to_csv('MTA_all_data_processed2.csv')

# read in zipcode data
zips = pd.read_csv("station_zips.csv")

# clean up of station names

def fix_names(name):
    temp_name = name.upper()
    replacements = [["1TH", "1"],["1ST", "1"], ["2ND", "2"],["3RD", "3"],["4TH", "4"],["5TH", "5"], 
                    ["6TH", "6"],["7TH", "7"],["8TH", "8"],["9TH", "9"],["0TH", "0"],["COLUMBIA UNIVERSITY", "COLUMBIA"],
                    ["AVENUE", "AV"], ["SQUARE", "SQ"], ["CENTER","CTR"], ["138 ST-3 AVE","3 AV 138 ST"],
                    ["149 ST-3 AV", "3 AV-149 ST"], ["137 ST-CITY COLLEGE", "137 ST CITY COL"], 
                    ["PROSPECT PARK-15 ST","15 ST-PROSPECT"], ['YANKEE STADIUM-161 ST', "161/YANKEE STAD"],
                    ['RAWSON ST-33 ST', "33 ST-RAWSON ST"], ['WHITEHALL ST', "WHITEHALL S-FRY"], 
                    ['WESTCHESTER SQ-EAST TREMONT AV',"WESTCHESTER SQ" ], ['WAKEFIELD-241 ST', "WAKEFIELD/241"],
                    ['WEST 8 ST', "W 8 ST-AQUARIUM"], ['WEST 4 ST', 'W 4 ST-WASH SQ'], 
                    ['VERNON BLVD-JACKSON AV',"VERNON-JACKSON" ],["5 AV-53 ST","5 AV/53 ST"],["5 AV","5 AVE"],
                    ["59 ST-COLUMBUS CIRCLE","59 ST COLUMBUS"],["WOODSIDE AV-61 ST","61 ST WOODSIDE"],
                    ["63 DRIVE-REGO PARK","63 DR-REGO PARK"],["66 ST-LINCOLN CTR","66 ST-LINCOLN"],
                    ["68 ST-HUNTER COLLEGE","68ST-HUNTER CO"],["AQUEDUCT-NORTH CONDUIT AV","AQUEDUCT N.COND"],
                    ["AQUEDUCT RACETRACK","AQUEDUCT RACETR"],["ASTOR PLACE","ASTOR PL"],
                    ["ASTORIA BLVD-HOYT AV","ASTORIA BLVD"],["DITMARS BLVD","ASTORIA DITMARS"],
                    ["ATLANTIC AV-BARCLAYS CTR","ATL AV-BARCLAY"],["BROADWAY-LAFAYETTE ST","B'WAY-LAFAYETTE"],
                    ["BAY PARKWAY","BAY PKWY"],["BEDFORD PARK BLVD","BEDFORD PK BLVD"],
                    ["BEDFORD-NOSTRAND AVS","BEDFORD-NOSTRAN"],["BEVERLY RD","BEVERLEY ROAD"],
                    ["BOTANIC GARDENS","BOTANIC GARDEN"],["BRIARWOOD-VAN WYCK BLVD","BRIARWOOD"],
                    ["BROADWAY JUNCTION-EAST NEW YORK","BROADWAY JCT"],["BROOKLYN BRIDGE-CITY HALL","BROOKLYN BRIDGE"],
                    ["CANARSIE - ROCKAWAY PARKWAY","CANARSIE-ROCKAW"],["CATHEDRAL PARKWAY-110 ST","CATHEDRAL PKWY"],
                    ["CLINTON-WASHINGTON AVS","CLINTON-WASH AV"],["DELANCEY ST","DELANCEY/ESSEX"],
                    ["EAST 143 ST-ST MARY'S ST","E 143/ST MARY'S"],["EAST 149 ST","E 149 ST"],
                    ["EAST 180 ST","E 180 ST"],["EASTCHESTER-DYRE AV","EASTCHSTER/DYRE"],
                    ["EASTERN PARKWAY-BROOKLYN MUSEUM","EASTN PKWY-MUSM"],["HARLEM-148 ST","HARLEM 148 ST"],
                    ["FAR ROCKAWAY-MOTT AV","FAR ROCKAWAY"],["FLATBUSH AV-BROOKLYN COLLEGE","FLATBUSH AV-B.C"],
                    ["FLUSHING-MAIN ST","FLUSHING-MAIN"],["FOREST AV","FOREST AVE"],
                    ["FOREST HILLS-71 AV","FOREST HILLS 71"],["FORT HAMILTON PARKWAY","FT HAMILTON PKY"],
                    ["GRAND ARMY PLAZA","GRAND ARMY PLAZ"],["GRAND AV-NEWTOWN","GRAND-NEWTOWN"],
                    ["HOWARD BEACH","HOWARD BCH JFK"],["HOYT & SCHERMERHORN","HOYT-SCHER"],
                    ["HUNTERS POINT","HUNTERS PT AV"],["INWOOD - 207 ST","INWOOD-207 ST"],
                    ["JAMAICA-179 ST","JAMAICA 179 ST"],["JACKSON HEIGHTS-ROOSEVELT AVE","JKSN HT-ROOSVLT"],
                    ["JAMAICA-VAN WYCK","JAMAICA VAN WK"],["KEW GARDENS-UNION TURNPIKE","KEW GARDENS"],
                    ["KINGS HIGHWAY","KINGS HWY"],["KNICKERBOCKER AV","KNICKERBOCKER"],["KOSCIUSKO ST","KOSCIUSZKO ST"],
                    ["NORWOOD-205 ST","NORWOOD 205 ST"],["OCEAN PARKWAY","OCEAN PKWY"],["PELHAM PARKWAY","PELHAM PKWY"],
                    ["QUEENSBORO PLAZA","QUEENSBORO PLZ"],["PARKCHESTER-EAST 177 ST","PARKCHESTER"],
                    ["LEXINGTON AV-53 ST","LEXINGTON AV/53"],["LEXINGTON AV","LEXINGTON AV/63"],
                    ["MARBLE HILL-225 ST","MARBLE HILL-225"],["METS - WILLETS POINT","METS-WILLETS PT"],
                    ["MORRISON AV-SOUNDVIEW AV","MORISN AV/SNDVW"],["MOSHOLU PARKWAY","MOSHOLU PKWY"],
                    ["MYRTLE-WILLOUGHBY AVS","MYRTLE-WILLOUGH"],["MYRTLE AV","MYRTLE-WYCKOFF"],
                    ["NEPTUNE AV-VAN SICLEN","NEPTUNE AV"],["238 ST-NEREID AV","NEREID AV"],
                    ["ROCKAWAY PARK-BEACH 116","ROCKAWAY PARK B"],["ROOSEVELT ISLAND","ROOSEVELT ISLND"],
                    ["SENECA AV","SENECA AVE"],["SUTPHIN BLVD-ARCHER AV - JFK","SUTPHIN-ARCHER"],
                    ["SUTTER AV","SUTTER AV-RUTLD"],["VAN CORTLANDT PARK-242 ST","V.CORTLANDT PK"],
                    ["VAN SICLEN AV","VAN SICLEN AVE"], ["AV H",'AVENUE H'], ["AV I",'AVENUE I'], ["AV J",'AVENUE J'],
                    ["AV M",'AVENUE M'], ["AV N",'AVENUE N'], ["AV P",'AVENUE P'], ["AV U",'AVENUE U'], 
                    ["AV X",'AVENUE X']]
    
       
    for x in replacements:
        temp_name = temp_name.replace(x[0], x[1])
    return temp_name

zips["STATION"] = zips["STATION"].apply(fix_names)

# merge zip code data with processed turnstile data
zipped_station = pd.merge(turnstile_df[[ 'UNIT','DIVISION', 'STATION', 'DATE', 'TIME','DATETIME', 'DAY', 
                                        'ENTRIES_DELTA', 'EXITS_DELTA']], zips[["DIVISION","STATION","zip"]],
                          on=["DIVISION","STATION"], how="left")


# read in file to map zip codes to csv
manual_zip_df = pd.read_csv("name_zip_mapping.csv")

# fixing/cleaning remaining station names
def fix_manual_names(cols):
    station_name= cols[0]
    zipcode=cols[1]
    if np.isnan(zipcode):
        values = manual_zip_df[manual_zip_df["name"]==station_name]["zip"].values
        if len(values)>0:
            return values[0]
        else:
            return np.nan
    else:
        return zipcode
        

# applying the cleaning function to zip codes and mapping stations to zip codes
zipped_station["zip"] = zipped_station[["STATION", "zip"]].apply(fix_manual_names,  axis=1)

read_dictionary = np.load('zipcode_to_income_dict.npy').item()

# read in income data
def add_income(zip_code):
    if zip in read_dictionary.keys():
        return read_dictionary[zip_code]


zipped_station["income"] = zipped_station["zip"].apply(add_income)

zipped_station.sample(30)
